[#11409] refactor(idp-basic): Inline IdP config validation into IdpRESTFeature
Remove the standalone validator class per review feedback and keep the
same fail-fast behavior with tests moved to TestIdpRESTFeature.
